Update patch set 1
I would prefer that you didn't submit this.

Lets put this change (and its related submissions) on hold at the
moment. For complete support of Thai, and other complex scripts, we are
planning to use ICU's layout library, which also handles compound
accents and placements. That will (I believe) supplant what this change
is trying to accomplish. Does that sound reasonable?

Patch-set: 1