Update patch set 1
Patch Set 1:

Hi Jean
Thank you for the help and the email. Please find my comments below:

Patch Set 1: Fails

Thanks :)

Lokesh>> Ops, we will work on this to make this through. We did compile everything here twice before and after committing though.

Note #1: We (Google) have been tightening the build system in our internal branches, in a way that now requires LOCAL_MODULE_TAGS := optional to be added to every new module. You'll need to add such a line right after LOCAL_MODULE:= libfemft in your Android.mk. From that point you'll need to add libfemft in build/target/product/core.mk
Lokesh>> We have fixed this. We are checking in the changes.

Note #2: My local editor complains that most newly added lines are CR-LF-terminated while the existing lines are LF-terminated, and it'd be good to fix that if possible.
Lokesh>> I am sorry for the same. We are fixing the same. For some internal review work etc, the file got contaminated with Windows Style line termination. We have fixed this and will checkin the same soon.

I'm also finding that the emulator fails to start in Google's internal master once this change and its companion are applied, but I'm not sure what the root cause is.
Lokesh>> This is strange to us. We tried to reproduce the same and could not. It worked very fine for us, even on Mac. If possible can you please do a complete cleanup and build. For us its working very fine here.


Regards
Lokesh

Patch-set: 1
Reviewer: Lokesh Joshi <1003708@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0