Update patch set 2
Patch Set 2: Do not submit

(1 inline comment)

Overall, I think we do not have to submit this.
It's too overdoing for the purpose.
Also, many code paths and priorities are changed (encoding, font sets).

And use "git commit -a --amend" if you update this CL. (Do not abandon and re-upload. It's very confusing.)

But anyway, I have a negative opinion on submitting this.

Patch-set: 2