Update patch set 1
Patch Set 1:

Waiting for build results at the technical level.

On the non-technical front, I'm worried that this removes the NOTICE and MODULE_LICENSE files, and might be interpreted as a change of license from ASL2 to BSD, which is only OK if we can confirm that this is an exact copy of upstream code.

Patch-set: 1
Label: Code-Review=0