Update patch set 1
Patch Set 1:

The license terms of Skia have changed upstream, see the URL below. This CL only forks from the upstream codebase in order to fix compile or test case errors. It is the intention of this code base to be a mirror of the upstream repository, but there are times we do diverge for specific reasons.

http://code.google.com/p/skia/source/detail?spec=svn3125&r=1982

Patch-set: 1
Label: Code-Review=0
Label: Verified=0