Update patch set 1
Patch Set 1:

As far as I can remember, I measured with skia bench. In some cases I had to modify it to use the shader/blitter I wanted to do measurements on.

Some of the tests combines several shaders and blitters, so that had to be taken into account as well.

As for the 'network problem', I don't have a clue. It's unlikely that it's one of the shaders or blitters, though.

Patch-set: 1