Update patch set 1
Patch Set 1:

To benchmark, run skia_bench with '-scale -forceFilter 1 -repeat 10'.

The improvement is visible in subtest repeatTile_index8.
On a Galaxy Nexus@1.2GHz the subtest is improved by 51% (From 292ms to 193ms).

When factoring in the overhead from the test by measuring the time of an empty shader implementation and subtracting it (13ms) from the result, the actual shader performance improvement was 55%.

Patch-set: 1
Reviewer: Henrik Smiding <1017245@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0