Update patch set 8
Patch Set 8:

The short answer is that yes, you can build a static lib that turns on that flag and then just include that static lib. The long answer is that you have to be very careful doing that.  Most importantly, you need this to ensure that the CFLAGS are identical between your static lib and the main skia library.  

Can you upload a CL that demonstrates your proposed approach?

Patch-set: 8
Label: Verified=0