Update patch set 15
Patch Set 15:

The fix to enable the *SSSE3.cpp file to be built as part of the framework was checked into Skia. It will be merged into the Android tree and be included in an upcoming release.  However, if you need it available in AOSP now, feel free to include those changes in this CL.

https://code.google.com/p/skia/source/detail?r=13583

Patch-set: 15