Update patch set 1
Patch Set 1:

For the emulator __SSSE3__ won't be defined. So:

#if !defined(SK_BUILD_FOR_ANDROID_FRAMEWORK) || SK_CPU_SSE_LEVEL >= SK_CPU_SSE_LEVEL_SSSE3

equals:

#if 0

The issue for the case with getcpuid(...) is that SK_BUILD_FOR_ANDROID_FRAMEWORK being defined will mean we don't compile in SSSE3 support. Line 45 of:

https://android.googlesource.com/platform/external/skia/+/master/include/core/SkUserConfig.h

I believe the better fix is to replace occurances of:

#if !defined(SK_BUILD_FOR_ANDROID_FRAMEWORK) || SK_CPU_SSE_LEVEL >= SK_CPU_SSE_LEVEL_SSSE3

with:

#if 1 || SK_CPU_SSE_LEVEL >= SK_CPU_SSE_LEVEL_SSSE3

which is:

#if 1

ie just remove the #if and always have SSSE3 support.

Patch-set: 1