Update patch set 1
Patch Set 1:

Ok, I believe I see the issue clearly now.  The problem case is when we compile the emulator without SSSE support, but then run on a device that supports SSSE3.  You're right that having a minimum requirement of SSSE3 for the emulator would solve the issue, but I'm not in a position to make that call.  Do you know if there is any defines set that allow us to determine if we are being run in an emulator?

Patch-set: 1