Update patch set 1
Patch Set 1: Code-Review+2

So I talked to Ian about this and we are ok with taking the change until we have time to roll a better solution.  I'll follow up with a CL that compiles the SSSE3 file in its own static lib that is included by skia so it can always be compiled with SSSE3.

Patch-set: 1
Reviewer: Derek Sollenberger <1006692@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2