Update patch set 1
Patch Set 1: Code-Review-1

This is something that should be in SkPostConfig.h not in each individual src file.  This has already been fixed in the main Skia repo and will land in AOSP on a future public push, but if you want to cherry-pick that feature now then grab lines 296-315 (from the file below) and replace what is current in that file.

https://skia.googlesource.com/skia/+/master/include/core/SkPostConfig.h

Patch-set: 1
Reviewer: Derek Sollenberger <1006692@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1