Update patch set 2
Patch Set 2:

We are currently discussing how to handle vendor specific replacements to the image codecs.  The current proposal is not viable with the new decoders we are working on, so we need to come up with a solution that we could support longer term.

Patch-set: 2
Reviewer: Derek Sollenberger <1006692@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0