Update patch set 1
Patch Set 1:

As a bit of background, we're working on a project that deals with a few benchmarking suites. In order for Skia's nanobench to produce meaningful results for us, we need to apply this patch. Given that it's already been accepted upstream, it would be convenient if we didn't have to carry it until whenever AOSP updates Skia (...which seems to be a rare occurrence). :)

Patch-set: 1
Label: Verified=0