Update patch set 2
Patch Set 2:

DO NOT MERGE doesn't work, because it needs to merge through some internal branches but not all.  Merged-In with the change-id of the internal change it conflicts with should prevent the automerger from merging it, but treehugger is not honoring it correctly (b/33746272), which is why it failed treehugger.

Patch-set: 2
Reviewer: Colin Cross <1002751@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0