Update patch set 1
Patch Set 1: Code-Review-1

Updating a global tolerance for over a thousand tests is too broad a change for something that only affects a small set of tests.  If we are to relax the restriction we should be able to do it over a much smaller set of test cases.

Patch-set: 1
Reviewer: Gerrit User 1006692 <1006692@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1