Update patch set 2
Patch Set 2:

The merge conflicts are because presubmit tries to apply this change in internal master, which has other conflicting changes from upstream Skia.

The appropriate thing to do here is to mention the Change-ID of the CL when this merged to master (http://ag/3908673) like below:
Merged-In: I0a36cec8829aee0081072b25e838096f99602cb3

Then presubmit will not test it in internal master (and automerger will skip applying the merge to any branch that already has http://ag/3908673.

Patch-set: 2
CC: Gerrit User 1058860 <1058860@85c56323-6fa9-3386-8a01-6480fb634889>