Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Brian, any motivation to have the rect-version *always* make a copy, even if it is the same size as the surface... as a signal that the caller doesn't want copy-on-write behavior?

Yes, I think that makes more sense - although that seems like a discussion for a fix to Skia ToT? And from the comments on the linked bug, are we still planning to land this? Or is there a newer/better fix coming?

Patch-set: 2
Reviewer: Gerrit User 1111168 <1111168@85c56323-6fa9-3386-8a01-6480fb634889>