Update patch set 1
Patch Set 1:

> Patch Set 1: Presubmit-Ready+1
> 
> > Patch Set 1:
> > 
> > third_party/gif has already been removed from Skia and will be removed from this repo when the next release is pushed.  Will we need to update the metadata again when that push happens?
> 
> Yes. If you know all the details of the upstream repository, a better option would be for you to create a full METADATA file like Dan did in http://aosp/1299576 and have our copy kept up to date automatically. If you did that, I would just abandon this change.

Google manages the upstream repository (skia.googlesource.com) and it active development is in support of Google projects.  Looking at that CL it seems to reference the specific revision the repository is pinned to and other details that wouldn't make as much sense in our case.

For now I think this is a sufficient approach.

Patch-set: 1
Reviewer: Gerrit User 1006692 <1006692@85c56323-6fa9-3386-8a01-6480fb634889>